Finishes mocking and cleaning of 3d-nurbs test
